[Tizen] Match the arguments type in 'max' function
- Fix the matching:
  'max(double, float)' -> 'max(double, double)'

+++ dali-toolkit/public-api/controls/scrollable/scroll-view/scroll-view.cpp
@@ -248,7 +248,7 @@ unsigned int FixedRuler::GetPageFromPosition(float position, bool wrap) const
     {
       position = WrapInDomain(position, mDomain.min, mDomain.max);
     }
-    page = std::max(static_cast<double>(0.0f), floor((position - mDomain.min) / mSpacing + 0.5f));
+    page = std::max(static_cast<double>(0.0f), static_cast<double>(floor((position - mDomain.min) / mSpacing + 0.5f)));
 
     if(wrap)
     {
